requirements.txt
streamlit
pandas
python-docx
PyYAML
streamlit-authenticator
python-dotenv
